+++ rtic-macros/src/codegen/bindings/esp32c3.rs
@@ -242,8 +242,7 @@ mod esp32c3 {
         stmts
     }
 
-pub fn extra_modules(_app: &App, _analysis: &SyntaxAnalysis) -> Vec<TokenStream2> {
-    vec![]
+    pub fn extra_modules(_app: &App, _analysis: &SyntaxAnalysis) -> Vec<TokenStream2> {
+        vec![]
+    }
 }
-}
-